requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
apify-client==1.12.1
google-auth==2.29.0
google-api-python-client==2.122.0
gspread==6.0.2
pydantic==2.11.7
google-generativeai==0.8.5
